+++ nrf_802154/driver/src/nrf_802154_trx.h
@@ -196,7 +196,8 @@ void nrf_802154_trx_channel_set(uint8_t channel);
 /**@brief Updates CCA configuration in the RADIO peripheral according to PIB. */
 void nrf_802154_trx_cca_configuration_update(void);
 
-/**@brief Puts the trx module into receive frame mode.
+/**
+ * @brief Put the trx module into receive frame mode.
  *
  * The frame will be received into buffer set by @ref nrf_802154_trx_receive_buffer_set.
  *
@@ -222,12 +223,10 @@ void nrf_802154_trx_cca_configuration_update(void);
  *                  responsibility to prepare the stimulation of this (D)PPI.
  * @param[in] notifications_mask Selects additional notifications generated during a frame reception.
  *                  It is bitwise combination of @ref nrf_802154_trx_receive_notifications_t values.
- * @param[in] p_ack_tx_power Selects the power which should be used to transmitted an ACK if required.
  */
-void nrf_802154_trx_receive_frame(uint8_t                                 bcc,
-                                  nrf_802154_trx_ramp_up_trigger_mode_t   rampup_trigg_mode,
-                                  nrf_802154_trx_receive_notifications_t  notifications_mask,
-                                  const nrf_802154_fal_tx_power_split_t * p_ack_tx_power);
+void nrf_802154_trx_receive_frame(uint8_t                                bcc,
+                                  nrf_802154_trx_ramp_up_trigger_mode_t  rampup_trigg_mode,
+                                  nrf_802154_trx_receive_notifications_t notifications_mask);
 
 /**@brief Puts the trx module into receive ACK mode.
  *
@@ -321,7 +320,8 @@ bool nrf_802154_trx_receive_is_buffer_missing(void);
  */
 bool nrf_802154_trx_receive_buffer_set(void * p_receive_buffer);
 
-/**@brief Begins frame transmit operation.
+/**
+ * @brief Begin frame transmit operation.
  *
  * This operation performs differently according to cca_attempts parameter.
  * When cca_attempts==0:
@@ -366,7 +366,7 @@ bool nrf_802154_trx_receive_buffer_set(void * p_receive_buffer);
  *                           transmission before the medium is considered busy. If 0, no CCA will be
  *                           performed. Otherwise, CCA procedures will be performed back to back until
  *                           idle channel is detected or @p cca_attempts attempts detect busy channel.
- * @param p_tx_power         Transmit power in dBm.
+ * @param p_tx_power         Pointer to the structure holding TX power split into raw components in dBm.
  * @param notifications_mask Selects additional notifications generated during a frame transmission.
  *                           It is bitwise combination of @ref nrf_802154_trx_transmit_notifications_t values.
  * @note To transmit ack after frame is received use @ref nrf_802154_trx_transmit_ack.
@@ -377,7 +377,8 @@ void nrf_802154_trx_transmit_frame(const void                            * p_tra
                                    const nrf_802154_fal_tx_power_split_t * p_tx_power,
                                    nrf_802154_trx_transmit_notifications_t notifications_mask);
 
-/**@brief Puts the trx module into transmit ACK mode.
+/**
+ * @brief Put the trx module into transmit ACK mode.
  *
  * @note This function may be called from @ref nrf_802154_trx_receive_frame_received handler only.
  *       This is because in this condition only the TIMER peripheral is running and allows timed transmission.
@@ -385,6 +386,7 @@ void nrf_802154_trx_transmit_frame(const void                            * p_tra
  * @param[in] p_transmit_buffer     Pointer to a buffer containing ACK frame to be transmitted.
  *                                  Caller is responsible for preparing an ACK frame according to the 802.15.4 protocol.
  * @param[in] delay_us              Delay (in microseconds)
+ * @param[in] p_tx_power_split      Pointer to the structure holding TX power split into raw components in dBm.
  *
  * @retval true     If the function was called in time and ACK frame is scheduled for transmission.
  *                  When transmission starts the function @ref nrf_802154_trx_transmit_ack_started will be called.
@@ -395,7 +397,9 @@ void nrf_802154_trx_transmit_frame(const void                            * p_tra
  *                  The TIMER peripheral is stopped and it is not possible to trigger @ref nrf_802154_trx_transmit_ack
  *                  again without receiving another frame again. No handlers will be called.
  */
-bool nrf_802154_trx_transmit_ack(const void * p_transmit_buffer, uint32_t delay_us);
+bool nrf_802154_trx_transmit_ack(const void                            * p_transmit_buffer,
+                                 uint32_t                                delay_us,
+                                 const nrf_802154_fal_tx_power_split_t * p_tx_power_split);
 
 /**@brief Puts trx module into IDLE mode.
  *
@@ -414,9 +418,10 @@ void nrf_802154_trx_standalone_cca(void);
 
 #if NRF_802154_CARRIER_FUNCTIONS_ENABLED
 
-/**@brief Starts generating continuous carrier.
+/**
+ * @brief Start generating continuous carrier.
  *
- * @param[in] p_tx_power  Transmit power in dBm.
+ * @param[in] p_tx_power  Pointer to the structure holding TX power split into raw components in dBm.
  *
  * Generation of a continuous carrier generates no handlers. It may be terminated by a call to
  * @ref nrf_802154_trx_abort or @ref nrf_802154_trx_disable.
@@ -433,10 +438,11 @@ void nrf_802154_trx_continuous_carrier(const nrf_802154_fal_tx_power_split_t * p
  */
 void nrf_802154_trx_continuous_carrier_restart(void);
 
-/**@brief Starts generating modulated carrier with given buffer.
+/**
+ * @brief Start generating modulated carrier with given buffer.
  *
  * @param[in] p_transmit_buffer Pointer to a buffer used for modulating the carrier wave.
- * @param[in] p_tx_power        Transmit power in dBm.
+ * @param[in] p_tx_power        Pointer to the structure holding TX power split into raw components in dBm.
  */
 void nrf_802154_trx_modulated_carrier(const void                            * p_transmit_buffer,
                                       const nrf_802154_fal_tx_power_split_t * p_tx_power);